+++ CHANGELOG.md
@@ -85,6 +85,8 @@ This section may not be all-inclusive; sorry!  I _did_ warn you that
 - Similarly, make `ShapeColorBuffers` not store GPU objects, so it can be
   constructed independently through `ShapeColorBuffers::new` (and shared between
   threads).
+- Rename the workspace buffer objects from `Buffers` to `Workspace` (e.g.
+  `fidget::wgpu::voxel::Buffers`) to more clearly reflect their usage.
 
 # 0.5.0
 This is a large release with a bunch of small features, reorganization, and one